Initial commit of file
Odyssey HMS macro to insert a note into each student in the selected
file.  File should be a text file of ID numbers, one per line.  Create a
new macro in the UI, add this file as contents, move it into proper menu
area and it should work just fine.